fix: hotfixed the hw and fw bugs. hw: add 2 trigger reg to start dma fifo write, and supporting 64bit addr, but still 16B aligned dma transferring; fw: update dma fifo reg writing as hw updating, and now it can support the plddr(63GB+1GB management space), which is included a xsa in `xsa/qkv_acc_top_v1.0.xsa`

+++ fw/nvme/nvme_identify.c
@@ -56,7 +56,7 @@
 #include "nvme.h"
 #include "nvme_identify.h"
 
-void controller_identification(unsigned int pBuffer)
+void controller_identification(unsigned long long pBuffer)
 {
 	ADMIN_IDENTIFY_CONTROLLER *identifyCNTL;
 	ADMIN_IDENTIFY_POWER_STATE_DESCRIPTOR *powerStateDesc;
@@ -147,7 +147,7 @@ void controller_identification(unsigned int pBuffer)
 	powerStateDesc->RWL = 0x0;
 }
 
-void namespace_identification(unsigned int pBuffer)
+void namespace_identification(unsigned long long pBuffer)
 {
 	ADMIN_IDENTIFY_NAMESPACE *identifyNS;
 	ADMIN_IDENTIFY_FORMAT_DATA *formatData;
